added support for setting presentation texts (#6034)
* added support for setting presentation texts

* updated project version

* removed code smells

* added test app

* added test case for instantiation

* current progress

* completed tests for presentation texts on instnace

* updated calculation test

* removed code smells

* added unittest

* fixed build error

* fixed failing test

* Update Skjema.cs

* Update DataApiTest.cs

* newest storage interface

* getting presentation texts from helper

* moved datahelper to new project

* no possible to update full instance object

* Assigning instance.presentationTexts from response from service

* removed code smells from testdata

* added unittest for service

* removed unused function

* no longer retrieving old data element

* return new dictionary if presentation texts not set

* fixed bug in datahelper

* using ToString instead of casting

* added null value handling

* removed double arrange

* fixed code smells in test project

* more specific exception type in dataHelper

+++ src/Altinn.Apps/AppTemplates/AspNet/App.IntegrationTests/Utils/TestDataUtil.cs
@@ -5,7 +5,7 @@ using App.IntegrationTests.Mocks.Services;
 
 namespace App.IntegrationTests.Utils
 {
-    public class TestDataUtil
+    public static class TestDataUtil
     {
         public static void PrepareInstance(string org, string app, int instanceOwnerId, Guid instanceGuid)
         {
@@ -14,6 +14,25 @@ namespace App.IntegrationTests.Utils
             string preInstancePath = instancePath.Replace(".json", ".pretest.json");
 
             File.Copy(preInstancePath, instancePath, true);
+
+            string dataPath = GetDataPath(org, app, instanceOwnerId, instanceGuid);
+
+            if (Directory.Exists(dataPath))
+            {
+                foreach (string filePath in Directory.GetFiles(dataPath, "*.*", SearchOption.AllDirectories))
+                {
+                    if (filePath.Contains(".pretest.json"))
+                    {
+                        // Handling all data elements
+                        File.Copy(filePath, filePath.Replace(".pretest.json", ".json"), true);
+                    }
+                    else if (filePath.EndsWith(".pretest"))
+                    {
+                        // Handling all data blobs
+                        File.Copy(filePath, filePath.Replace(".pretest", string.Empty), true);
+                    }
+                }
+            }
         }
 
         public static void DeleteInstance(string org, string app, int instanceOwnerId, Guid instanceGuid)
@@ -39,28 +58,34 @@ namespace App.IntegrationTests.Utils
         public static void DeleteDataForInstance(string org, string app, int instanceOwnerId, Guid instanceGuid)
         {
             string path = GetDataPath(org, app, instanceOwnerId, instanceGuid);
+
             if (Directory.Exists(path))
             {
-                Directory.Delete(path, true);
+                foreach (string filePath in Directory.GetFiles(path, "*.*", SearchOption.AllDirectories))
+                {
+                    if (!filePath.Contains("pretest"))
+                    {
+                        File.Delete(filePath);
+                    }
+                }
+
+                if (Directory.GetFiles(path).Length == 0)
+                {
+                    Directory.Delete(path, true);
+                }
             }
         }
 
         private static string GetInstancePath(string org, string app, int instanceOwnerId, Guid instanceGuid)
         {
-            string unitTestFolder = Path.GetDirectoryName(new Uri(typeof(InstanceMockSI).Assembly.CodeBase).LocalPath);
+            string unitTestFolder = Path.GetDirectoryName(new Uri(typeof(InstanceMockSI).Assembly.Location).LocalPath);
             return Path.Combine(unitTestFolder, @"..\..\..\Data\Instances\", org + @"\", app + @"\", instanceOwnerId + @"\", instanceGuid.ToString() + @".json");
         }
 
         private static string GetDataPath(string org, string app, int instanceOwnerId, Guid instanceGuid)
         {
-            string unitTestFolder = Path.GetDirectoryName(new Uri(typeof(InstanceMockSI).Assembly.CodeBase).LocalPath);
+            string unitTestFolder = Path.GetDirectoryName(new Uri(typeof(InstanceMockSI).Assembly.Location).LocalPath);
             return Path.Combine(unitTestFolder, @"..\..\..\Data\Instances\", org + @"\", app + @"\", instanceOwnerId + @"\", instanceGuid.ToString() + @"\");
-        }
-
-        private static string GetDataBlobPath(string org, string app, int instanceOwnerId, Guid instanceGuid, Guid dataId)
-        {
-            string unitTestFolder = Path.GetDirectoryName(new Uri(typeof(InstanceMockSI).Assembly.CodeBase).LocalPath);
-            return Path.Combine(unitTestFolder, @"..\..\..\Data\Instances\", org + @"\", app + @"\", instanceOwnerId + @"\", instanceGuid.ToString() + @"\blob\" + dataId.ToString());
         }
     }
 }
